Fix Volume Mapping with Docker Sync
Seems like when `volume` is being used in `docker-compose.yml` the one defined
in `docker-compose.override.yml` is not being honored. To fix this I moved
all volumes mapping to override files. Now there are 2 types of override files
which are for `linux` and `mac`. The mac will use Docker Sync and linux will use
local driver.