Sync Apps Folder
**Why we need this change?**

Previously, `dockery-sync` only sync one application. This is troublesome
because each time you need to change to other application, you will need
to run `resync.sh` script which will take few minutes to complete.

With current change, `dockery-sync` will sync applications parent folder.
Basically, we are syncing all the applications folders.
So to change to other application, even a new project we just created,
we don't need to run `resync.sh` again.
The downside of this is, we are mounting all the applications to the container.